Converted to using Jsonal for file I/O.
Jsonal is *very* fast.  A 70 MB file parses in around 1/3 of a second (on a very good machine).

Error reporting should also be slightly better if the JSON in the WCON files has problems.

Warning: this depends on `kse.jsonal` which is _not yet published_.  As soon as it is, this branch can be merged.  Do not merge yet.