Added degrees and radians as units to Scala reader/writer.
Default unit is radians (better for math).

Allowed variants are "deg", "degree", "degrees", "rad", "radian", "radians" with the 3-character forms taking the short SI prefixes (e.g. "mrad") and the others taking the long form (e.g. "centidegrees").

Documentation is not updated to reflect these as this is a proposal rather than a final decision on how to support angular units.

Note that supporting revolutions would be easy also at least in the Scala code base.