Unit conversion tests.
Most contain multiple files where all values after conversion should be the
same, and all values before conversion should be the same when the filenames
are the same save for the (optional) digit before the extension.

In addition, there is a test to make sure conversions work inside custom
annotated blocks.